refactor: extract ApiResult component from ApiResults
